feat(table/reflow): reject tables with invalid separator rows during reflow
- Changed format_separator_cells to return empty vector if length mismatch, so invalid separator rows are detected.
- Added handling in calculate_and_format to return None when separator formatting is invalid, preventing table reflow.
- Made reflow_table return original input lines if separator row is invalid, preserving original table.
- Refactored tests by moving them out of reflow.rs into reflow/tests.rs with expanded coverage.
- Minor docs fix in ADR title case.

This prevents malformed Markdown tables with invalid separator rows from being reflowed incorrectly, improving stability and correctness of the table reflow pipeline.
