+++ docs/adrs/0001-table-reflow-pipeline.md
@@ -1,4 +1,4 @@
-# Architecture Decision Record (ADR) 0001: preserve table structure during reflow
+# Architecture Decision Record (ADR) 0001: Preserve table structure during reflow
 
 - Status: Accepted
 - Date: 2026-04-22
